+++ apis/models/requests.py
@@ -73,11 +73,11 @@ class CommonRequest(BaseModel):
     input_image_checkbox: bool = Field(default=False, description="Input Image")
     current_tab: StrictStr = Field(default='uov', description="Current tab")
     uov_method: UpscaleOrVaryMethod = Field(default=UpscaleOrVaryMethod.disable, description="Upscale or Vary Method")
-    uov_input_image: str | None = Field(default=None, description="Upscale or Vary Input Image")
+    uov_input_image: StrictStr | None = Field(default="None", description="Upscale or Vary Input Image")
     outpaint_selections: List[OutpaintExpansion] = Field(default=[], description="Outpaint Expansion")
-    inpaint_input_image: str | None = Field(default=None, description="Inpaint Input Image")
+    inpaint_input_image: StrictStr | None = Field(default="None", description="Inpaint Input Image")
     inpaint_additional_prompt: StrictStr = Field(default="", description="Additional prompt for inpaint")
-    inpaint_mask_image_upload: str | None = Field(default=None, description="Inpaint Mask Image Upload")
+    inpaint_mask_image_upload: str | None = Field(default="None", description="Inpaint Mask Image Upload")
     disable_preview: bool = Field(default=False, description="Disable preview")
     disable_intermediate_results: bool = Field(default=False, description="Disable intermediate results")
     disable_seed_increment: bool = Field(default=False, description="Disable seed increment")
