+++ apis/utils/call_worker.py
@@ -2,24 +2,24 @@
 Calls the worker with the given params.
 """
 import asyncio
-import copy
 import os
 import json
 import uuid
 import datetime
-from pathlib import Path
 
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from fastapi import Response
+
+from apis.models.base import CurrentTask
 from apis.models.response import RecordResponse
 from apis.utils.api_utils import params_to_params
-from apis.utils.file_utils import save_base64
+from apis.utils.pre_process import pre_worker
 from apis.utils.sql_client import GenerateRecord
+from apis.utils.post_worker import post_worker, url_path
 
 from apis.utils.img_utils import (
-    narray_to_base64img,
-    base64_from_path
+    narray_to_base64img
 )
 from apis.models.requests import CommonRequest
 from modules.async_worker import AsyncTask, async_tasks
@@ -38,77 +38,10 @@ Session = sessionmaker(bind=engine, autoflush=True)
 session = Session()
 
 
-class CurrentTask:
-    """
-    Current task class.
-    """
-    ct = None
-
-
-# todo: use argument to specify hosts
-def url_path(result: list) -> list:
-    """
-    Converts the result to a list of URL paths.
-    :param result: The result to convert.
-    :return: The list of URL paths.
-    """
-    url_or_path = []
-    if len(result) == 0:
-        return url_or_path
-    if str.startswith(result[0], 'http'):
-        for res in result:
-            uri = '/'.join(res.split('/')[-2:])
-            url_or_path.append(os.path.join(OUT_PATH, uri))
-        return url_or_path
-    for res in result:
-        path = Path(res).as_posix()
-        uri = '/'.join(path.split('/')[-2:])
-        url_or_path.append(f"http://127.0.0.1:7866/outputs/{uri}")
-    return url_or_path
-
-
-def pre_worker(request: CommonRequest):
-    """
-    Pre-processes the request.
-    :param request: The request to pre-process.
-    :return: The pre-processed request.
-    """
-    os.makedirs(INPUT_PATH, exist_ok=True)
-    req_copy = copy.deepcopy(request)
-    req_copy.inpaint_input_image = save_base64(req_copy.inpaint_input_image, INPUT_PATH)
-    cn_imgs = []
-    for cn in req_copy.controlnet_image:
-        cn.cn_img = save_base64(cn.cn_img, INPUT_PATH)
-        cn_imgs.append(cn)
-    req_copy.controlnet_image = cn_imgs
-    return req_copy
-
-
-def post_worker(task: AsyncTask, started_at: int):
-    """
-    Posts the task to the worker.
-    :param task: The task to post.
-    :param started_at: The time the task started.
-    :return: The task.
-    """
-    try:
-        query = session.query(GenerateRecord).filter(GenerateRecord.task_id == task.task_id).first()
-        query.start_mills = started_at
-        query.finish_mills = int(datetime.datetime.now().timestamp() * 1000)
-        query.task_status = "finished"
-        query.progress = 100
-        query.result = url_path(task.results)
-        session.commit()
-    except Exception as e:
-        print(e)
-    CurrentTask.ct = None
-    return task
-
-
 async def execute_in_background(task: AsyncTask, raw_req: CommonRequest, in_queue_mills):
     """
     Executes the request in the background.
-    :param request: The request to execute.
+    :param task: The task to execute.
     :param raw_req: The raw request.
     :param in_queue_mills: The time the request was enqueued.
     :return: The response.
@@ -147,14 +80,14 @@ async def stream_output(request: CommonRequest):
     Calls the worker with the given params.
     :param request: The request object containing the params.
     """
-    raw_req = pre_worker(request)
+    raw_req, request = await pre_worker(request)
     params = params_to_params(request)
     task = AsyncTask(args=params)
     async_tasks.append(task)
-    in_queue_mills=int(datetime.datetime.now().timestamp() * 1000)
+    in_queue_mills = int(datetime.datetime.now().timestamp() * 1000)
     session.add(GenerateRecord(
         task_id=task.task_id,
-        req_params=raw_req.model_dump_json(),
+        req_params=json.loads(raw_req.model_dump_json()),
         webhook_url=raw_req.webhook_url,
         in_queue_mills=in_queue_mills
     ))
@@ -163,7 +96,7 @@ async def stream_output(request: CommonRequest):
     started = False
     finished = False
     while not finished:
-        await asyncio.sleep(0.01)
+        await asyncio.sleep(0.2)
         if len(task.yields) > 0:
             if not started:
                 started = True
@@ -213,14 +146,14 @@ async def binary_output(request: CommonRequest):
     :param request: The request object containing the params.
     """
     request.image_number = 1
-    raw_req = pre_worker(request)
+    raw_req, request = await pre_worker(request)
     params = params_to_params(request)
     task = AsyncTask(args=params, task_id=uuid.uuid4().hex)
     async_tasks.append(task)
-    in_queue_mills=int(datetime.datetime.now().timestamp() * 1000)
+    in_queue_mills = int(datetime.datetime.now().timestamp() * 1000)
     session.add(GenerateRecord(
         task_id=task.task_id,
-        req_params=raw_req.model_dump_json(),
+        req_params=json.loads(raw_req.model_dump_json()),
         webhook_url=raw_req.webhook_url,
         in_queue_mills=in_queue_mills
     ))
@@ -229,7 +162,7 @@ async def binary_output(request: CommonRequest):
     started = False
     finished = False
     while not finished:
-        await asyncio.sleep(0.01)
+        await asyncio.sleep(0.2)
         if len(task.yields) > 0:
             if not started:
                 started = True
@@ -255,17 +188,17 @@ async def binary_output(request: CommonRequest):
                 finished = True
                 post_worker(task=task, started_at=started_at)
 
-    with open (task.results[0], "rb") as f:
+    with open(task.results[0], "rb") as f:
         image = f.read()
     return Response(image, media_type="image/png")
 
 
-async def async_worker(request: CommonRequest) -> RecordResponse:
+async def async_worker(request: CommonRequest) -> dict:
     """
     Calls the worker with the given params.
     :param request: The request object containing the params.
     """
-    raw_req = pre_worker(request)
+    raw_req, request = await pre_worker(request)
     task_id = uuid.uuid4().hex
     task = AsyncTask(
         task_id=task_id,
@@ -275,7 +208,7 @@ async def async_worker(request: CommonRequest) -> RecordResponse:
     in_queue_mills = int(datetime.datetime.now().timestamp() * 1000)
     session.add(GenerateRecord(
         task_id=task.task_id,
-        req_params=raw_req.model_dump_json(),
+        req_params=json.loads(raw_req.model_dump_json()),
         webhook_url=raw_req.webhook_url,
         in_queue_mills=in_queue_mills
     ))
