+++ requirements_versions.txt
@@ -18,3 +18,4 @@ onnxruntime==1.16.3
 timm==0.9.2
 sse_starlette
 sqlalchemy
+aiohttp
